+++ src/hapi/pipelines/database/org_type.py
@@ -1,11 +1,11 @@
 import logging
 from typing import Dict
-from unicodedata import normalize
 
 from hapi_schema.db_org_type import DBOrgType
 from hdx.scraper.utilities.reader import Read
 from sqlalchemy.orm import Session
 
+from ..utilities.mappings import get_code_from_name
 from .base_uploader import BaseUploader
 
 logger = logging.getLogger(__name__)
@@ -41,14 +41,9 @@ class OrgType(BaseUploader):
         self._session.commit()
 
     def get_org_type_code(self, org_type: str) -> str:
-        # TODO: implement fuzzy matching of org types (HAPI-194)
-        org_type_code = self.data.get(org_type)
-        if org_type_code:
-            return org_type_code
-        org_type = (
-            normalize("NFKD", org_type)
-            .encode("ascii", "ignore")
-            .decode("ascii")
+        org_type_code = get_code_from_name(
+            name=org_type,
+            code_lookup=self.data,
+            code_mapping=self._org_type_map,
         )
-        org_type_code = self._org_type_map.get(org_type.lower())
         return org_type_code